requirements.txt
requests~=2.26
django~=2.2
djangorestframework~=3.12
PyJWT~=2.1
pytest~=6.2
pytest-django~=4.4
pytest-pythonpath~=0.7
